방생성관련 기능 생성 (DB에 관련 데이터 생성 및 Gemini API 연동, Solved.ac API 연동)

+++ app/src/main/java/com/example/softwareproject/data/dto/problem/PsProblemDto.kt
@@ -2,7 +2,7 @@ package com.example.softwareproject.data.dto.problem
 
 data class PsProblemDto(
     val acceptedUserCount: Int = 0,
-    val averageTries: Int = 0,
+    val averageTries: Double = 0.0,
     val codingRoomId: String = "",
     val problemId: String = "",
     val problemIndex: String = "",
